Split out message types
