+++ packages/@n8n/task-runner-python/src/message_types/__init__.py
@@ -1,0 +1,25 @@
+from .broker import (
+    BrokerMessage,
+    BrokerInfoRequest,
+    BrokerRunnerRegistered,
+    BrokerTaskOfferAccept,
+)
+from .runner import (
+    RunnerMessage,
+    RunnerInfo,
+    RunnerTaskOffer,
+    RunnerTaskAccepted,
+    RunnerTaskRejected,
+)
+
+__all__ = [
+    "BrokerMessage",
+    "BrokerInfoRequest",
+    "BrokerRunnerRegistered",
+    "BrokerTaskOfferAccept",
+    "RunnerMessage",
+    "RunnerInfo",
+    "RunnerTaskOffer",
+    "RunnerTaskAccepted",
+    "RunnerTaskRejected",
+]
