+++ packages/@n8n/task-runner-python/src/message_types/broker.py
@@ -1,0 +1,26 @@
+from dataclasses import dataclass
+from typing import Literal, Union
+
+
+@dataclass
+class BrokerInfoRequest:
+    type: Literal["broker:inforequest"] = "broker:inforequest"
+
+
+@dataclass
+class BrokerRunnerRegistered:
+    type: Literal["broker:runnerregistered"] = "broker:runnerregistered"
+
+
+@dataclass
+class BrokerTaskOfferAccept:
+    task_id: str
+    offer_id: str
+    type: Literal["broker:taskofferaccept"] = "broker:taskofferaccept"
+
+
+BrokerMessage = Union[
+    BrokerInfoRequest,
+    BrokerRunnerRegistered,
+    BrokerTaskOfferAccept,
+]
